split indexing modules -> file_converter + preprocessor
