+++ haystack/document_store/sql.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import relationship, sessionmaker
 
 from haystack.document_store.base import BaseDocumentStore
 from haystack import Document, Label
-from haystack.indexing.utils import eval_data_from_file
+from haystack.preprocessor.utils import eval_data_from_file
 
 Base = declarative_base()  # type: Any
