+++ rest_api/controller/file_upload.py
@@ -13,8 +13,8 @@ from rest_api.config import DB_HOST, DB_PORT, DB_USER, DB_PW, DB_INDEX, ES_CONN_
     SEARCH_FIELD_NAME, FILE_UPLOAD_PATH, EMBEDDING_DIM, EMBEDDING_FIELD_NAME, EXCLUDE_META_DATA_FIELDS, VALID_LANGUAGES, \
     FAQ_QUESTION_FIELD_NAME, REMOVE_NUMERIC_TABLES, REMOVE_WHITESPACE, REMOVE_EMPTY_LINES, REMOVE_HEADER_FOOTER
 from haystack.document_store.elasticsearch import ElasticsearchDocumentStore
-from haystack.indexing.file_converters.pdf import PDFToTextConverter
-from haystack.indexing.file_converters.txt import TextConverter
+from haystack.file_converter.pdf import PDFToTextConverter
+from haystack.file_converter.txt import TextConverter
 
 
 logger = logging.getLogger(__name__)
