+++ test/test_imports.py
@@ -1,8 +1,8 @@
 def test_module_imports():
     from haystack import Finder
     from haystack.document_store.sql import SQLDocumentStore
-    from haystack.indexing.cleaning import clean_wiki_text
-    from haystack.indexing.utils import convert_files_to_dicts, fetch_archive_from_http
+    from haystack.preprocessor.cleaning import clean_wiki_text
+    from haystack.preprocessor.utils import convert_files_to_dicts, fetch_archive_from_http
     from haystack.reader.farm import FARMReader
     from haystack.reader.transformers import TransformersReader
     from haystack.retriever.sparse import TfidfRetriever
